// frontend/src/interfaces/auth.ts
export interface UserRegisterPayload {
  name: string;
  email: string;
  password: string;
}

export interface UserLoginPayload {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  // accounts?: Account[];
}